Add AppNavigator in Example
